Add response codes to all API endpoints.
Added proper response codes and cleaned up the code of the controllers for the /users and /courses endpoints.
Added a ResourceNotFoundException, which represents a 404 Not found response code. The exception is a RuntimeException, made to be thrown whenever a resource doesn't exist in the database. Whenever it gets thrown from a method, a response gets returned with 404.
Cleaned up UserService and CourseService.

+++ StudentManagementSystem/src/main/java/com/yoanpetrov/studentmanagementsystem/controller/CourseController.java
@@ -1,13 +1,17 @@
 package com.yoanpetrov.studentmanagementsystem.controller;
 
+import com.yoanpetrov.studentmanagementsystem.exceptions.ResourceNotFoundException;
 import com.yoanpetrov.studentmanagementsystem.model.Course;
 import com.yoanpetrov.studentmanagementsystem.model.User;
 import com.yoanpetrov.studentmanagementsystem.service.CourseService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
+
+// TODO: 29-Sep-23 Add a service method to remove all of the course's users
 
 @RequiredArgsConstructor
 @RestController
@@ -17,57 +21,72 @@ public class CourseController {
     private final CourseService courseService;
 
     @GetMapping
-    public List<Course> getAllCourses() {
-        return courseService.getAllCourses();
+    public ResponseEntity<List<Course>> getAllCourses() {
+        List<Course> courses = new ArrayList<>(courseService.getAllCourses());
+        if (courses.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(courses, HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
-    public Optional<Course> getCourseById(@PathVariable Long id) {
-        return courseService.getCourseById(id);
+    public ResponseEntity<Course> getCourseById(@PathVariable Long id) {
+        Course course = courseService.getCourseById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Course not found"));
+        return new ResponseEntity<>(course, HttpStatus.OK);
     }
 
     @GetMapping("/{id}/users")
-    public List<User> getAllCourseUsers(@PathVariable Long id) {
-        return courseService.getAllCourseUsers(id);
+    public ResponseEntity<List<User>> getAllCourseUsers(@PathVariable Long id) {
+        if (!courseService.existsCourse(id)) {
+            throw new ResourceNotFoundException("Course not found");
+        }
+        List<User> users = courseService.getAllCourseUsers(id);
+        if (users.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(users, HttpStatus.OK);
     }
 
     @PostMapping
-    public Course createCourse(@RequestBody Course course) {
-        return courseService.createCourse(course);
+    public ResponseEntity<Course> createCourse(@RequestBody Course course) {
+        Course createdCourse = courseService.createCourse(course);
+        return new ResponseEntity<>(createdCourse, HttpStatus.CREATED);
     }
 
-    /**
-     * Adds a user to a course
-     *
-     * @param id the course id.
-     * @param requestUser the user to be added to the course.
-     * @return the added user todo everything has to change to use response codes and stuff.
-     */
-    @PostMapping("/{id}/users")
-    public User addUserToCourse(@PathVariable Long id, @RequestBody User requestUser) {
-        User user = courseService.addUserToCourse(id, requestUser);
-        return user;
+    @PostMapping("/{courseId}/users")
+    public ResponseEntity<User> addUserToCourse(@PathVariable Long courseId, @RequestBody User requestUser) {
+        User user = courseService.addUserToCourse(courseId, requestUser);
+        return new ResponseEntity<>(user, HttpStatus.CREATED);
     }
 
     @PutMapping("/{id}")
-    public Course updateCourse(@PathVariable Long id, @RequestBody Course courseDetails) {
-        return courseService.updateCourse(id, courseDetails);
+    public ResponseEntity<Course> updateCourse(@PathVariable Long id, @RequestBody Course courseDetails) {
+        if (!courseService.existsCourse(id)) {
+            throw new ResourceNotFoundException("Course not found");
+        }
+        Course course = courseService.updateCourse(id, courseDetails);
+        return new ResponseEntity<>(course, HttpStatus.OK);
     }
 
     @DeleteMapping
-    public String deleteAllCourses() {
+    public ResponseEntity<String> deleteAllCourses() {
         courseService.deleteAllCourses();
-        return "All users have been deleted successfully.";
+        return new ResponseEntity<>("All courses have been deleted successfully", HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteCourse(@PathVariable Long id) {
+    public ResponseEntity<String> deleteCourse(@PathVariable Long id) {
+        if (!courseService.existsCourse(id)) {
+            throw new ResourceNotFoundException("Course not found");
+        }
         courseService.deleteCourse(id);
+        return new ResponseEntity<>("Course deleted successfully", HttpStatus.OK);
     }
 
-    @DeleteMapping("/{id}/users")
-    public User removeUserFromCourse(@PathVariable Long id, @RequestBody User requestUser) {
-        User user = courseService.removeUserFromCourse(id, requestUser);
-        return user;
+    @DeleteMapping("/{courseId}/users")
+    public ResponseEntity<User> removeUserFromCourse(@PathVariable Long courseId, @RequestBody User requestUser) {
+        User user = courseService.removeUserFromCourse(courseId, requestUser);
+        return new ResponseEntity<>(user, HttpStatus.OK);
     }
 }
